+++ lib/floop.dart
@@ -1,6 +1,6 @@
 library floop;
 
-export './src/observed.dart' show floop, DynMap, DynValue;
+export './src/observed.dart' show floop, DynMap, DynValue, DynList;
 export './src/mixins.dart'
     show Floop, FloopWidget, FloopStateful, FloopStatefulWidget, DynamicWidget;
 export './src/repeater.dart' show Repeater;
@@ -10,5 +10,4 @@ export './src/transition.dart'
         transitionEval,
         transitionOf,
         TransitionGroup,
-        Transitions, // ignore: deprecated_member_use_from_same_package
         TransitionsConfig;